Extract Slide Design methods from GraphicsUtil
